Ch. 17 §02: correct numbering and content for Listings 17-10 and 17-11
