+++ listings/ch17-async-await/listing-17-11/src/main.rs
@@ -4,46 +4,22 @@ fn main() {
     trpl::block_on(async {
         let (tx, mut rx) = trpl::channel();
 
-        let tx1 = tx.clone();
-        let tx1_fut = async move {
-            let vals = vec![
-                String::from("hi"),
-                String::from("from"),
-                String::from("the"),
-                String::from("future"),
-            ];
+        let vals = vec![
+            String::from("hi"),
+            String::from("from"),
+            String::from("the"),
+            String::from("future"),
+        ];
 
-            for val in vals {
-                tx1.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        };
+        for val in vals {
+            tx.send(val).unwrap();
+            trpl::sleep(Duration::from_secs(1)).await;
+        }
 
-        let rx_fut = async {
-            while let Some(value) = rx.recv().await {
-                println!("received '{value}'");
-            }
-        };
-
-        let tx_fut = async move {
-            let vals = vec![
-                String::from("more"),
-                String::from("messages"),
-                String::from("for"),
-                String::from("you"),
-            ];
-
-            for val in vals {
-                tx.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        };
-
-        // ANCHOR: here
-        let futures =
-            vec![Box::new(tx1_fut), Box::new(rx_fut), Box::new(tx_fut)];
-
-        trpl::join_all(futures).await;
-        // ANCHOR_END: here
+        // ANCHOR: loop
+        while let Some(value) = rx.recv().await {
+            println!("received '{value}'");
+        }
+        // ANCHOR_END: loop
     });
 }
